feature-flag/03_openfeature_flagd/app/requirements.txt
flask==3.1.0
openfeature-sdk==0.7.5
openfeature-provider-flagd==0.4.2
